+++ src/app/(afterLogin)/_component/BlogCreateForm.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { useState, useRef } from "react";
 import Input from "@/app/_component/Input";
 import TextArea from "@/app/_component/TextArea";
 import Checkbox from "@/app/_component/Checkbox";
@@ -8,65 +7,22 @@ import SelectBox from "../_component/SelectBox";
 import SubmitButton from "@/app/_component/SubmitButton";
 import ImageUploader from "@/app/(afterLogin)/_component/ImageUploader";
 import { Categories } from "@/constants/data";
-import { useCreateBlog } from "@/hooks/useBlog";
-import { BlogFormData } from "@/types/blog";
-import { validateBlogForm } from "../_utils/validateBlog";
+import { useBlogForm } from "../_hooks/useBlogForm";
 
 export default function BlogCreateForm() {
-  const [mainImage, setMainImage] = useState<File | null>(null);
-  const [subImage, setSubImage] = useState<File | null>(null);
-  const [selectedCategory, setSelectedCategory] = useState<string>("");
-  const [isAgreed, setIsAgreed] = useState<boolean>(false);
-
-  const titleRef = useRef<HTMLInputElement>(null);
-  const contentRef = useRef<HTMLTextAreaElement>(null);
-
-  const { mutate: createBlog, isPending } = useCreateBlog();
-
-  const handleMainImageChange = (file: File | null) => {
-    setMainImage(file);
-  };
-
-  const handleSubImageChange = (file: File | null) => {
-    setSubImage(file);
-  };
-
-  const handleCategoryChange = (category: string) => {
-    setSelectedCategory(category);
-  };
-
-  const handleAgreementChange = (checked: boolean) => {
-    setIsAgreed(checked);
-  };
-
-  const handleSubmit = async () => {
-    // 이용 정책 동의 검사
-    if (!isAgreed) {
-      alert("Blog 이용 정책 동의가 필요합니다.");
-      return;
-    }
-
-    const title = titleRef.current?.value || "";
-    const content = contentRef.current?.value || "";
-
-    // 폼 데이터 구성
-    const formData: BlogFormData = {
-      title,
-      content,
-      mainImage,
-      subImage,
-      category: selectedCategory,
-    };
-
-    // 유효성 검사
-    const validationError = validateBlogForm(formData);
-    if (validationError) {
-      alert(validationError);
-      return;
-    }
-
-    createBlog(formData);
-  };
+  const {
+    mainImage,
+    selectedCategory,
+    isAgreed,
+    titleRef,
+    contentRef,
+    isPending,
+    handleMainImageChange,
+    handleSubImageChange,
+    handleCategoryChange,
+    handleAgreementChange,
+    handleSubmit,
+  } = useBlogForm();
 
   return (
     <div className="flex flex-col gap-6 mx-5">
